% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_path}
\alias{sim_path}
\title{Simulate a jump-process path}
\usage{
sim_path(i, rates, dists, t = 0, u = 0, tn = Inf, abs = numeric(0), bs = NA)
}
\arguments{
\item{i}{Initial state (integer).}

\item{rates}{Function `rates(state, time, duration)` returning total jump rate.}

\item{dists}{Function `dists(state, time, duration)` returning next-state probabilities.}

\item{t}{Initial time. Default is `0`.}

\item{u}{Initial duration since the last jump. Default is `0`.}

\item{tn}{Maximum simulation time. Default is `Inf`.}

\item{abs}{Logical vector indicating absorbing states. If omitted, the last
state is treated as absorbing.}

\item{bs}{Optional upper bounds for rates by state. If omitted, bounds are
approximated by optimization.}
}
\value{
A list with elements `times` and `states`.
}
\description{
Simulates the path of a time-inhomogeneous Markov or semi-Markov jump
process up to a maximum time.
}
\examples{
jump_rate <- function(i, t, u) if (i == 1) 2 + t else if (i == 2) 1 + t else 0
mark_dist <- function(i, s, v) if (i == 1) c(0, 0.7, 0.3) else if (i == 2) c(0, 0, 1) else 0
sim_path(1, rates = jump_rate, dists = mark_dist, tn = 2)
}
